// src/Client/TreeMon.NG/src/app/models/appinfo.ts
// Copyright 2015, 2017 TreeMon.org.
// Licensed under CPAL 1.0,  See license.txt  or go to http://treemon.org/docs/license.txt  for full license details.

export class AppInfo {

     ActiveDbProvider: string;

     ActiveDbConnectionKey: string;

     DatabaseServer: string;

     ActiveDatabase: string;

     ActiveDbUser: string;

     ActiveDbPassword: string;

     AppType: string; // forms, mobile

     AccountName: string;

     AccountEmail: string;

     AccountIsPrivate: boolean;

     UserName: string;

     UserEmail: string;

     UserIsPrivate: boolean;

     UserPassword: string;

    ConfirmPassword: string;

    SecurityQuestion: string;

     UserSecurityAnswer: string;

    PasswordSalt: string;

     RunInstaller: boolean;

     SiteDomain: string;

    Command: string;
}